backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
scikit-learn
shap
numpy
pandas
google-generativeai
python-dotenv
pydantic[email]
pydantic-settings